Changed the file tree a little bit, to make it neater / more like the boilerplate. I like the way that they do it -- it's much neater to put everything in folders, and then have an index.
ready to build the actions for making the api call for setting a user's location. I'm not sure if that's actually a bad idea -- does it make more sense to have one api call to update the entire user object? It seems like the best thing is to have a POST method that updates whatever part of the user method you want it to. that makes the most sense. Otherwise, you'd start writing paths for updating each little bit of the user's record, bit by bit: `updateName`, `updateSomeOtherPref`, `updateEmail`, and that's just silly.

The thing that I worry about is that we're getting certain information from twitter -- and that info we know is good. but I don't want to trust the client on everything. Or anything, really, not more than i have to. Which, looking at the user model that I set up, is pretty easy: the only information that we need from the user is their location. so an 'update all fields that we trust the user about' and 'update location' are the same thing.

Good talk, team.